+++ src/scriptlets/scriptlets-list.js
@@ -61,3 +61,4 @@ export * from './remove-node-text';
 export * from './trusted-replace-node-text';
 export * from './evaldata-prune';
 export * from './trusted-prune-inbound-object';
+export * from './spoof-css';